updated user 1 goal 4

+++ src/main/java/com/example/oop_project_group_8/Mahir_2211582/Order.java
@@ -10,7 +10,7 @@ public class Order {
     private String shipmentDetails;
 
     // Constructor
-    public Order(String orderId, String datePlaced, String timePlaced, float totalAmount, String paymentStatus, String shipmentStatus, String shipmentDetails) {
+    public Order(String orderId, String datePlaced, String timePlaced, double totalAmount, String paymentStatus, String shipmentStatus, String shipmentDetails) {
         this.orderId = orderId;
         this.datePlaced = datePlaced;
         this.timePlaced = timePlaced;
